Changed name of the cache adapter from Apc to APC
